requirements.txt
# Core
telethon>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Utilities
python-telegram-bot>=20.0
httpx>=0.25.0
apscheduler>=3.10.0

# AI/ML (optional)
openai>=1.0.0
anthropic>=0.18.0

# Logging
python-json-logger>=2.0.0

# Type hints
typing-extensions>=4.8.0
